Add proxy.config.http.max_proxy_cycles (#7076)
* Removed #include <string_view> added by prior commit.

* Added static_cast<int> for casting string_view::size_type to int.

+++ proxy/http/HttpTransact.cc
@@ -26,7 +26,6 @@
 
 #include <strings.h>
 #include <cmath>
-#include <string_view>
 
 #include "HttpTransact.h"
 #include "HttpTransactHeaders.h"
@@ -6594,7 +6593,7 @@ HttpTransact::will_this_request_self_loop(State *s)
         while ((count <= max_proxy_cycles) && (std::string_view::npos != (offset = current.find(uuid)))) {
           current.remove_prefix(offset + TS_UUID_STRING_LEN);
           count++;
-          TxnDebug("http_transact", "count = %d current = %.*s", count, current.length(), current.data());
+          TxnDebug("http_transact", "count = %d current = %.*s", count, static_cast<int>(current.length()), current.data());
         }
       }
